+++ R/h_survival_duration_subgroups.R
@@ -13,7 +13,6 @@
 #' @order 1
 #'
 #' @examples
-#'
 #' # Testing dataset.
 #' library(scda)
 #' library(dplyr)
@@ -39,21 +38,22 @@
 #'   )
 #' labels <- c("ARM" = adtte_labels[["ARM"]], "SEX" = adtte_labels[["SEX"]], "is_event" = "Event Flag")
 #' formatters::var_labels(adtte_f)[names(labels)] <- labels
+#'
+#' @keywords internal
 NULL
 
 #' @describeIn h_survival_duration_subgroups helper to prepare a data frame of median survival times by arm.
 #' @inheritParams h_survival_duration_subgroups
 #'
-#' @export
-#'
 #' @examples
-#'
 #' # Extract median survival time for one group.
-#' h_survtime_df(
+#' tern:::h_survtime_df(
 #'   tte = adtte_f$AVAL,
 #'   is_event = adtte_f$is_event,
 #'   arm = adtte_f$ARM
 #' )
+#'
+#' @keywords internal
 h_survtime_df <- function(tte, is_event, arm) {
   assertthat::assert_that(
     is.numeric(tte),
@@ -102,12 +102,10 @@ h_survtime_df <- function(tte, is_event, arm) {
 #'    in a data frame. `variables` corresponds to the names of variables found in `data`, passed as a named list and
 #'    requires elements `tte`, `is_event`, `arm` and optionally `subgroups`. `groups_lists` optionally specifies
 #'    groupings for `subgroups` variables.
-#' @export
 #'
 #' @examples
-#'
 #' # Extract median survival time for multiple groups.
-#' h_survtime_subgroups_df(
+#' tern:::h_survtime_subgroups_df(
 #'   variables = list(
 #'     tte = "AVAL",
 #'     is_event = "is_event",
@@ -118,7 +116,7 @@ h_survtime_df <- function(tte, is_event, arm) {
 #' )
 #'
 #' # Define groupings for BMRKR2 levels.
-#' h_survtime_subgroups_df(
+#' tern:::h_survtime_subgroups_df(
 #'   variables = list(
 #'     tte = "AVAL",
 #'     is_event = "is_event",
@@ -134,6 +132,8 @@ h_survtime_df <- function(tte, is_event, arm) {
 #'     )
 #'   )
 #' )
+#'
+#' @keywords internal
 h_survtime_subgroups_df <- function(variables,
                                     data,
                                     groups_lists = list(),
@@ -178,14 +178,15 @@ h_survtime_subgroups_df <- function(variables,
 #'   treatment hazard ratio.
 #' @param strata_data (`factor`, `data.frame` or `NULL`)\cr
 #'   required if stratified analysis is performed.
-#' @export
-#' @examples
 #'
+#' @examples
 #' # Extract hazard ratio for one group.
-#' h_coxph_df(adtte_f$AVAL, adtte_f$is_event, adtte_f$ARM)
+#' tern:::h_coxph_df(adtte_f$AVAL, adtte_f$is_event, adtte_f$ARM)
 #'
 #' # Extract hazard ratio for one group with stratification factor.
-#' h_coxph_df(adtte_f$AVAL, adtte_f$is_event, adtte_f$ARM, strata_data = adtte_f$STRATA1)
+#' tern:::h_coxph_df(adtte_f$AVAL, adtte_f$is_event, adtte_f$ARM, strata_data = adtte_f$STRATA1)
+#'
+#' @keywords internal
 h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control_coxph()) {
   assertthat::assert_that(
     is.numeric(tte),
@@ -285,11 +286,10 @@ h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control
 #'   `data`, passed as a named list and requires elements `tte`, `is_event`, `arm` and
 #'   optionally `subgroups` and `strat`. `groups_lists` optionally specifies
 #'   groupings for `subgroups` variables.
-#' @export
-#' @examples
 #'
+#' @examples
 #' # Extract hazard ratio for multiple groups.
-#' h_coxph_subgroups_df(
+#' tern:::h_coxph_subgroups_df(
 #'   variables = list(
 #'     tte = "AVAL",
 #'     is_event = "is_event",
@@ -300,7 +300,7 @@ h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control
 #' )
 #'
 #' # Define groupings of BMRKR2 levels.
-#' h_coxph_subgroups_df(
+#' tern:::h_coxph_subgroups_df(
 #'   variables = list(
 #'     tte = "AVAL",
 #'     is_event = "is_event",
@@ -318,7 +318,7 @@ h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control
 #' )
 #'
 #' # Extract hazard ratio for multiple groups with stratification factors.
-#' h_coxph_subgroups_df(
+#' tern:::h_coxph_subgroups_df(
 #'   variables = list(
 #'     tte = "AVAL",
 #'     is_event = "is_event",
@@ -328,6 +328,8 @@ h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control
 #'   ),
 #'   data = adtte_f
 #' )
+#'
+#' @keywords internal
 h_coxph_subgroups_df <- function(variables,
                                  data,
                                  groups_lists = list(),
@@ -399,10 +401,7 @@ h_coxph_subgroups_df <- function(variables,
 #'
 #' @return A list with subset data (`df`) and metadata about the subset (`df_labels`).
 #'
-#'
 #' @examples
-#'
-#'
 #' library(rtables)
 #'
 #' df <- data.frame(